requirements.txt
aiogram~=3.7
psutil~=5.9
python-dotenv~=1.0
